requirements.txt
PyJWT
requests
tweepy
boto3